This reader is kind of elegant, kind of fucked.
I don't know which to feel more strongly about yet. It is handy that it tracks
line and col numbers as you go.

But the entire calculation just passes a blackboard around and functions play
with what they want and leave the rest. I like it, but it worries me.

Time will tell, I suppose.