Working to the extent of nameless recursion.
If a y combinator evaluates properly then a lot of interpreter steps are lining
up right. Not a great test, but a start.

In more exciting news, my scheme of walking the ast to set dynamic bindings
breaks equality of identities since it changes every clojure.lang.IObj, so that
has to go. But switching the walk from (instance? clojure.lang.IObj to a
reasonable check grinds the entire system to a halt.