Ditched notion of contextswitches in favour of attached envs.
Immediates and Applications both represent computations "in progress" and so
have a state or local env. These are the natural points at which to save context
to be retrieved during future evaluation.

The problem now is that local environments collect cruft which makes it hard to
tell if two forms are equal without walking the whole tree and pruning the envs
based on symbols remaining below them (and in scope) in the tree.

That's not something I want to tackle just yet.