Allow def to delay if the name parameter cannot be evaluated yet.
This is working except for one big caveat: the env being extended is captured
lexically at present as a hack until we have statefuls. The problem, of course,
is that if a def is delayed, then when it later gets fulfilled, we lose anything
defined in the interim.

def needs to send an update event on the env channel, not a full environment.