Endless tweaking.
This impl is going the same way as the rest. I've managed to keep it compact,
but it's still the biggest one in a while and has more corners than a koch
snowflake.

There's something subtly flawed about the env management. I really need to
crystalise the ideas so that I can try to write something obvious.