Archival commit.
I think this entire approach is demonstrably a failure.

The basic idea of context switches is sound, but the mechanics just have so many
corner cases that I'm constantly losing track of what I intended.

I'm thinking of going back to my very first idea that symbols should be bound at
read time with a bit of extra tree walking to unset and set parameters.

As I've simplified the interpreter, approaches that used to fail have started to
work, so either I understand the problem better now and can solve it, or I'm
about to waste another week or two on a dead end I've trekked up before.