Select can walk all branches if p doesn't resolve.
That's conjecture, based on the observation that if p doesn't resolve then the
branches can't resolve any of the "gate parameters" in p either and so it's safe
to walk.

Not walking the branches lames the compiler and makes understanding the code
much more difficult, so I'm not without bias in the handwaving proof.

But I'll stick with it until I have reason not to.