Radical! Back to notation compatible with the blackboard.
Simplifying the env impl to match what I do by hand has both simplified the env
management and put me back exactly where I was on the first shot, but with a
different error message.

I was hopeing the rewrite would solve my problems, but it's still a kind of progress.