Added non-applicative if* construct.
This is sort of a hack to allow extra optimisations when you really know what
you're doing. But it also allows you to control when (and if) the branches of
branching contruct get evaluated.

The optimisation aspect should be automated eventually. We'll see if we still
need the construct after that. But it's so trivial to implement, taking it out
is at best a signal.