Replace syms with uniqued syms in μ bodies.
Back to autogensyms which still feels like a hack, but I've got a case now (fn)
where there's no other way to distinguish uses of the same parameter by
different μs.

Plus I've figured out a fairly minimal and robust way to implement it.

REVIEW: Should I create a type for uniqued syms so that we can still display
code as it was written? Or is it enough to simply annotate ns level definitions
with the original source code?